allTasks without scrollView
Display a brief description of the post in the post cell. Height
The camera must be dynamic. If the text will engage
is less than 2 lines, or will be equal to 2 lines, then
display all text by default (no height restrictions
text). If the same text takes up more than two lines, it needs to be put
ellipsis at the end of the line, cut off this text, set
restriction to respect only the first two lines, and
a displayed "Expand" button that will remove
limit text height to 2 lines and the cell will be
Fully display a short description of the post. Also in
In this case, the button should have the opposite state -
"Collapse" which will respect this constraint
again.